Moved away from cthulhu-character-generator module.
All the content from cthulhu-character-generator has now been moved to this repository as the original module would require a lot more work instead of moving the stuff just here.. so, here we go. We might drop them back into the old module later on and see how to handle it properly. But now I just want to make PROGRESS with the speed of a nimble ninja cat.